% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_rx.R
\name{id2key}
\alias{id2key}
\title{Indices to Factor}
\usage{
id2key(x, collapse = ";", none = "[none]", ...)
}
\arguments{
\item{x}{\link[base]{logical} \link[base]{matrix}}

\item{collapse}{\link[base]{character} scalar, symbol to
collapse all keywords}

\item{none}{\link[base]{character} scalar,
label for all-negative, default \code{'[none]'}}

\item{...}{..}
}
\value{
Function \link{id2key} returns a \link[base]{character} \link[base]{vector}.
Note that function \link{id2key} does \emph{not} return a \link[base]{factor},
as the levels will be very messy and may create problems when
\link[base]{rbind}.
}
\description{
..
}
\details{
..
}
\examples{
(x = cbind(
 a = c(1, 1, 1, 0, NA, 1, NA),
 b = c(NA, 0, 1, 0, NA, NA, NA),
 c = c(1, 0, 0, 0, 0, NA, NA)
))
storage.mode(x) = 'logical'
id2key(x)
}
\keyword{internal}
