% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_rx.R
\name{key_rx}
\alias{key_rx}
\title{Index to Factor, selected using Regular Expression}
\usage{
key_rx(pattern, envir = parent.frame(), ...)
}
\arguments{
\item{pattern, envir}{see function \link{select_rx}}

\item{...}{additional parameters of function \link{id2key}}
}
\description{
..
}
\details{
Function \link{key_rx} is the inverse of function \link{splitKey}.
}
\examples{
npk |> 
 within.data.frame(expr = {
  N = N |> as.character() |> as.numeric()
  P = P |> as.character() |> as.numeric()
  K = K |> as.character() |> as.numeric()
  tmp = key_rx(pattern = 'N|P|K')
 })
}
