% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_rx.R
\name{op_rx}
\alias{op_rx}
\alias{select_rx}
\alias{apply_rx}
\title{Operation on Objects Specified by \link[base]{regex}}
\usage{
select_rx(pattern, envir = parent.frame(), ...)

apply_rx(pattern, envir = parent.frame(), ..., FUN, MoreArgs = NULL)
}
\arguments{
\item{pattern}{\link[base]{character} scalar, regular expression}

\item{envir}{an \link[base]{environment},
a \link[base]{list},
a \link[base]{data.frame}.
May also be a \link[base]{matrix} for function \link{select_rx}.}

\item{...}{additional parameters of \link[base]{grep}, most importantly \code{invert}}

\item{FUN}{\link[base]{function}}

\item{MoreArgs}{(optional) \link[base]{list}, additional parameters of \code{FUN}}
}
\value{
Function \code{\link[=select_rx]{select_rx()}} selects

\itemize{
\item {columns from a \link[base]{data.frame} by \code{pattern} matching against its \link[base]{names},
and returns a \link[base]{data.frame}.}
\item {elements from a \link[base]{list} by \code{pattern} matching against its \link[base]{names},
and returns a \link[base]{list}.}
\item {columns from a \link[base]{matrix} by \code{pattern} matching against its \link[base]{colnames},
and returns a \link[base]{matrix}.}
\item {objects in an \link[base]{environment} by \code{pattern} matching against the names of variables within,
and returns a \strong{\link[base]{list}}
instead of an \link[base]{environment}.}
}

Function \code{\link[=apply_rx]{apply_rx()}} returns the updated \code{envir}, whether it is
\link[base]{environment}, \link[base]{list} or \link[base]{data.frame}.
}
\description{
..
}
\examples{
iris |> select_rx(pattern = '\\\\.Width$') |> head()
iris |> select_rx(pattern = '\\\\.Length$', invert = TRUE) |> head()
VADeaths |> select_rx(pattern = '^Rural')
iris |> head() |> with(expr = select_rx(pattern = '\\\\.Width$'))
attenu |> 
 head() |> 
 apply_rx(pattern = '^acc', FUN = stats:::format_perc, MoreArgs = list(digits = 2))

attenu |> 
 head() |> 
 within.data.frame(expr = {
  apply_rx('^acc', FUN = stats:::format_perc, MoreArgs = list(digits = 2))
 }) # same
}
\keyword{internal}
