% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitKey.R
\name{split_ext}
\alias{split_ext}
\alias{splitKey}
\alias{splitOrd}
\title{Split \link[base]{character} \link[base]{vector}}
\usage{
splitKey(x, key = xkey, data.name = substitute(x), envir = parent.frame(), ...)

splitOrd(
  x,
  nm = stop("must specify new names"),
  data.name = substitute(x),
  envir = parent.frame(),
  ...
)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}, each element being a set of
keywords separated by a symbol (e.g., \code{','})}

\item{key}{(optional for function \link{splitKey}) \link[base]{character} \link[base]{vector},
user-specified keywords.  Default to all keywords appearing in input \code{x}}

\item{data.name}{(optional) \link[base]{character} scalar or \link[base]{name},
name of input \code{x}}

\item{envir}{\code{NULL}, \code{FALSE} or an \link[base]{environment}.}

\item{...}{potential parameters of \link[base]{strsplit}, most importantly \code{split}}

\item{nm}{(for function \link{splitOrd}) \link[base]{character} \link[base]{vector}}
}
\value{
Function \code{\link[=splitKey]{splitKey()}} returns a \link[base]{logical} \link[base]{matrix} if \code{envir = NULL}.
Otherwise the \link[base]{logical} \link[base]{vector}s are assigned to \code{envir}
(i.e., when used inside \link[base]{within.data.frame}).

Function \link{splitOrd} returns a \link[base]{character} \link[base]{matrix} if \code{envir = NULL}.
Otherwise the \link[base]{character} \link[base]{vector}s are assigned to \code{envir}
(i.e., when used inside \link[base]{within.data.frame}).
}
\description{
Split \link[base]{character} \link[base]{vector}, into keywords or by the order of appearance.
}
\details{
Function \link{splitKey} finds out whether each keyword appears in each element of input \code{x}.

\code{NA_character_} or \code{''} entries in input \code{x} are regarded
as negative (i.e., none of the keywords exists),
instead of as missingness (i.e., we do not know if any of the keywords exists).
This practice is most intuitive to clinicians.
}
\examples{
letters[1:4] |> 
 splitKey(split = ';;', envir = NULL) # exception

(x1 = c('a,b,', 'c,a,b,,a', NA_character_, ''))
x1 |> 
 splitKey(split = ',', envir = NULL)

data.frame(x = x1) |> 
 within.data.frame(expr = splitKey(x, split = ','))

data.frame(x = x1) |> 
 within.data.frame(expr = splitKey(x, split = ',', data.name = 'cancer'))
\dontrun{
X = rep(x1, times = 10L)
microbenchmark::microbenchmark( # speed O(n)
 splitKey(x1, split = ',', envir = NULL), 
 splitKey(X, split = ',', envir = NULL))}
(x2 = c('T2;N0;M0;B1', '; ;M1; ', NA_character_, ''))
nm = c('T', 'N', 'M', 'B')
x2 |>
 splitOrd(split = ';', nm = nm, envir = NULL)
data.frame(x = x2) |> 
 within.data.frame(expr = splitOrd(x, split = ';', nm = nm, data.name = 'st'))
}
\keyword{internal}
